Move more net functionality into `humility-net`
