+++ humility-cli/Cargo.toml
@@ -13,7 +13,7 @@ parse_int.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 
-humility-net-core.workspace = true
+humility-net.workspace = true
 humility-probes-core = { workspace = true, optional = true }
 humility.workspace = true
 hubtools.workspace = true
