+++ humility-core/src/lib.rs
@@ -5,7 +5,6 @@ pub mod archive;
 pub mod core;
 pub mod dump;
 pub mod hubris;
-pub mod net;
 pub mod reflect;
 
 // Re-export `humility_log`, so clients don't need to import it (or `slog`)
